mode

+++ src/main/webapp/WEB-INF/jsp/usr/article/detail.jsp
@@ -1,15 +1,15 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
-	pageEncoding="UTF-8"%>
+    pageEncoding="UTF-8"%>
 <%@ taglib uri="jakarta.tags.core" prefix="c"%>
 
 <c:set var="pageTitle" value="상세보기" />
 
-<%@ include file="/WEB-INF/jsp/common/header.jsp"%>
+<%@ include file="/WEB-INF/jsp/common/header.jsp" %>
 
 <section class="mt-8">
 	<div class="container mx-auto">
-		<div class="table-box">
-			<table>
+		<div class="w-9/12 mx-auto">
+			<table class="table table-lg">
 				<tr>
 					<th>번호</th>
 					<td>${article.getId() }</td>
@@ -36,20 +36,19 @@
 				</tr>
 			</table>
 		</div>
-
-		<div class="btns mt-3 text-sm flex justify-between">
-			<button onclick="history.back();">뒤로가기</button>
-
-
-	
-				<c:if test="${rq.getLoginedMemberId() == article.getMemberId() }">
-					<button onclick="history.back();">뒤로가기</button>
-					
-					<a onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;"
-						href="doDelete?id=${article.getId() }">삭제</a>
-				</c:if>
+		
+		<div class="w-9/12 mx-auto mt-3 text-sm flex justify-between">
+			<div>	
+				<button class="btn btn-active btn-sm" onclick="history.back();">뒤로가기</button>
 			</div>
+			<c:if test="${rq.getLoginedMemberId() == article.getMemberId() }">
+				<div>
+					<a class="btn btn-active btn-sm" href="modify?id=${article.getId() }">수정</a>
+					<a class="btn btn-active btn-sm" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;" href="doDelete?id=${article.getId() }">삭제</a>
+				</div>
+			</c:if>
 		</div>
+	</div>
 </section>
 
-<%@ include file="/WEB-INF/jsp/common/footer.jsp"%>
+<%@ include file="/WEB-INF/jsp/common/footer.jsp" %>
